Family_Expenses_Management-backend/requirements.txt
fastapi>=0.100.0
uvicorn==0.22.0
pydantic==2.7.1
motor==3.3.1
pymongo==4.5
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
email-validator>=2.0
slowapi==0.1.9
apscheduler==3.10.1
PyJWT ==2.10.1
pydantic-settings ==  2.7.1
python-multipart == 0.0.20